@@ -10,7 +10,10 @@ import webpackMiddleware from "webpack-dev-middleware";
 import webpackHotMiddleware from "webpack-hot-middleware";
 import http from "http";
 import webpackConfig from "../webpack/dev.babel";
-import Config from "config";
+import Config from "../src/config";
+import {
+  isomorphicDevelopment
+} from "../settings";
 
 // Inform developers that Compilation is in process
 // and wait till its fully done
@@ -32,7 +35,7 @@ const commonClientMiddlewareInstance = webpackMiddleware(commonClientCompiler, {
   contentBase: commonClientConfig.devServer.contentBase,
   publicPath: commonClientConfig.output.publicPath,
   serverSideRender: true,
-  // Turn off the server-side rendering mode. See Server-Side Rendering part for more info.
+  // Turn on the server-side rendering mode. See Server-Side Rendering part for more info.
 });
 
 // Use the webpack middleware
@@ -71,23 +74,28 @@ if (serviceWorkerConfig !== null) {
     );
     
     const serviceWorkerContents = serviceWorkerMiddlewareInstance.fileSystem.readFileSync(`${serviceWorkerConfig.output.path}/service-worker.js`, "utf-8");
-    const swVersionHash = crypto.createHash("md5").update(serviceWorkerContents).digest("hex");
-    // eslint-disable-next-line
+    let swResponseText = `
+      var ASSETS = ${JSON.stringify(assets)};
+      ${serviceWorkerContents}
+    `;
+    const swVersionHash = crypto.createHash("md5").update(swResponseText).digest("hex");
+    swResponseText = `
+      var VERSION = "${swVersionHash}";
+      ${swResponseText}
+    `;
+    
     res.setHeader("Content-Type", "application/javascript");
     // No cache header
     res.setHeader("Cache-Control", "private, no-cache, no-store, must-revalidate");
     res.setHeader("Expires", "-1");
     res.setHeader("Pragma", "no-cache");
-    res.send(`
-      var VERSION = "${swVersionHash}";
-      var ASSETS = ${JSON.stringify(assets)};
-      ${serviceWorkerContents}
-    `);
+    res.send(swResponseText);
   });
 }
 
 // server content from content base
 app.use("/public", express.static(commonClientConfig.devServer.contentBase));
+
 // Add assets to request
 app.use(function (req, res, next) {
   let webpackStats = res.locals.webpackStats.toJson();
@@ -102,12 +110,11 @@ app.use(function (req, res, next) {
   _.each(webpackStats, stat => {
     const assetsByChunkName = stat.assetsByChunkName;
     const publicPath = stat.publicPath;
-
+    
     _.each(assetsByChunkName, (chunkValue, chunkName) => {
 
       // If its array then it just contains chunk value as array
       if (_.isArray(chunkValue)) {
-        //console.log(chunkValue);
         _.each(chunkValue, (path, index) => {
           assetsByChunkName[chunkName][index] = `${publicPath}${path}`;
         });
@@ -117,43 +124,48 @@ app.use(function (req, res, next) {
             assetsByChunkName[chunkName][subChunkType][subChunkIndex] = `${publicPath}${subChunkValue}`;
           });
         });
+      } else if (_.isString(chunkValue)) {
+        assetsByChunkName[chunkName] = `${publicPath}${chunkValue}`;
       }
     });
     assets = {...assets, ...assetsByChunkName};
   });
-  
   req.assets = assets;
   next();
 });
 
 // Include server routes as a middleware
 app.use(function(req, res, next) {
-  require("../src/server")(req, res, next);
+  // We need to require this run time!
+  require("../src/server").default(req, res, next);
 });
 
-const watcher = chokidar.watch("../src/server");
 
-watcher.on("ready", function() {
-  watcher.on("all", function() {
-    // eslint-disable-next-line
-    console.log("Clearing /src/ module cache from server");
-    Object.keys(require.cache).forEach(function(id) {
-      // eslint-disable-next-line
-      if (/[\/\\]src[\/\\]/.test(id)) delete require.cache[id];
-    });
+const clearRequireCache = () => {
+  // eslint-disable-next-line
+  console.log("Clearing module cache...");
+  Object.keys(require.cache).forEach(function(id) {
+    delete require.cache[id];
   });
-});
+};
+
+if (process.env.NODE_ENV === "development" && isomorphicDevelopment) {
+  
+  const watcher = chokidar.watch("../src");
+  watcher.on("ready", function() {
+    watcher.on("all", clearRequireCache);
+  });
+}
+
 
 
 // Do "hot-reloading" of react stuff on the server
 // Throw away the cached client modules and let them be re-required next time
 commonClientCompiler.plugin("done", function() {
-  // eslint-disable-next-line
-  console.log("Clearing /src/ module cache from server");
-  Object.keys(require.cache).forEach(function(id) {
-    // eslint-disable-next-line
-    if (/[\/\\]src[\/\\]/.test(id)) delete require.cache[id];
-  });
+  
+  if (process.env.NODE_ENV === "development" && isomorphicDevelopment) {
+    clearRequireCache();
+  }
   
   const addr = server.address();
   // eslint-disable-next-line
