@@ -7,7 +7,7 @@ import path from "path";
  *    images  --> server images
  *    public --> Public folder that can be pushed to cdn
  *      build --> Compiled files
- *    server.js --> Single compiled server file
+ *    middleware.js --> Single compiled server file
  *  node_modules
  *  src --> Main sources
  *    app
@@ -22,8 +22,8 @@ import path from "path";
  *    utils --> Util functions
  *    client.js --> Client entry js
  *    routes.js --> Consolidated routes
- *    server.js --> Server instance
- *    start-server.js --> Simple start server
+ *    middleware.js --> Server instance
+ *    start-middleware.js --> Simple start server
  */
 
 /**
@@ -31,7 +31,7 @@ import path from "path";
  * application building and in src folder
  * @type {string}
  */
-const publicDirName = "public";
+export const publicDirName = "public";
 
 /**
  * @description Distribution directory name where all the code will
@@ -86,3 +86,60 @@ export const srcPublicDir = path.resolve(path.join(srcDir, publicDirName));
 export const distPublicDir = path.join(distDir, publicDirName);
 
 export const buildDir = path.join(distPublicDir, buildDirName);
+
+export const images = {
+  useWebP: true,
+  allowedExtensions: [
+    ".bmp",
+    ".gif",
+    ".jpeg",
+    ".jpg",
+    ".png",
+    ".svg",
+  ],
+};
+
+/**
+ * Enabling this option will create a new file common-vendor which would contain
+ * all the 3rdp party modules used anywhere.
+ *
+ * The drawback of turning this setting ON is when number of node_modules increases
+ * the common-vendor JS file size increases
+ * The benefit of using this setting is, other code splitting is comparatively less
+ * so individual page JS file size decreases
+ *
+ * I prefer to let it stay ON and only turn it OFF when we need to use application in where
+ * bandwidth & large file size is not a concern
+ *
+ * @type {boolean}
+ */
+export const isolateVendorScripts = false;
+
+/**
+ * Enable service worker for application
+ * @type {boolean}
+ */
+export const enableServiceWorker = false;
+
+/**
+ * Isomorphic (Server rendering is disabled when developing app i.e. NODE_ENV = development) by default
+ * This settings help you manipulate the setting, if server side rendering should be cached or not!
+ *
+ * Setting it to true might cause severe performance issues
+ */
+export const isomorphicDevelopment = false;
+
+// export everything in single object
+export default {
+  directories: {
+    buildPublicPath,
+    rootDir,
+    distDir,
+    srcDir,
+    srcPublicDir,
+    distPublicDir,
+    buildDir
+  },
+  images,
+  isolateVendorScripts,
+};
