@@ -3,6 +3,7 @@ import PropTypes from "prop-types";
 import _ from "lodash";
 import { generateStringHash } from "../utils/utils";
 import { generateMeta} from "../utils/seo";
+import {enableServiceWorker} from "../../../settings";
 
 export default class Html extends React.Component {
   
@@ -33,21 +34,19 @@ export default class Html extends React.Component {
         <head>
           <title>{this.getTitle()}</title>
           {/** The url /manifest.json is a url handled via the server.js **/}
-          <link rel="manifest" href={"/manifest.json"}/>
+          {
+            enableServiceWorker &&
+            (<link rel="manifest" href={"/manifest.json"} />)
+          }
+          
           {
             _.map(this.getMeta(), (meta, i) => {
               return <meta key={i} {...meta} />;
             })
           }
-          <script dangerouslySetInnerHTML={{
-            __html: `(function(){var WebP=new Image();WebP.onload=WebP.onerror=function(){
-              if(WebP.height!=2){var sc=document.createElement('script');sc.type='text/javascript';sc.async=true;
-              var s=document.getElementsByTagName('script')[0];sc.src='public/js/webpjs-0.0.2.min.js';s.parentNode.insertBefore(sc,s);}};
-              WebP.src='data:image/webp;base64,UklGRjoAAABXRUJQVlA4IC4AAACyAgCdASoCAAIALmk0mk0iIiIiIgBoSygABc6WWgAA/veff/0PP8bA//LwYAAA';})();`
-          }} />
         </head>
         <body>
-          <div id="app">{this.props.children}</div>
+          <div id="app" dangerouslySetInnerHTML={{ __html: this.props.children}} />
           {
             _.map(scripts, path => {
               const pathHash = generateStringHash(path, "JS");
