@@ -1,9 +1,10 @@
 import React from "react";
 import { Provider } from "react-redux";
 import { render as DefaultRender } from "react-dom";
-import ConnectedRouter from "core/libs/ConnectedRouter";
 import { AppContainer as HotAppContainer } from "react-hot-loader";
 import _ from "lodash";
+
+import ConnectedRouter from "../libs/ConnectedRouter";
 
 import {
   Router as DefaultRouter,
@@ -11,35 +12,53 @@ import {
   Switch as DefaultSwitch,
 } from "react-router-dom";
 
-import clientStorage from "core/libs/storage/storage.server";
-import clientApi from "core/libs/api/api";
+import clientStorage from "../libs/storage";
+import clientApi from "../libs/api";
 
-import Loader from "core/components/loader/loader";
-import NotFoundPage from "app/components/error/404";
-import ErrorPage from "app/components/error/500";
-import RouteWithSubRoutes from "core/components/route/with-sub-routes";
+
+import RouteWithSubRoutes from "../components/route/with-sub-routes";
 
 import { getRouteFromPath } from "./bundler";
+import { getComponent } from "./utils";
+
+const RootComponent = getComponent("root");
+const Loader  = getComponent("loader");
+const NotFoundPage = getComponent("error/404");
+const ErrorPage = getComponent("error/500");
+const OfflinePage = getComponent("error/offline");
 
 export const renderNotFoundPage = ({
   render = DefaultRender,
   Router = DefaultRouter,
   Switch = DefaultSwitch,
   Route = DefaultRoute,
+  api = clientApi,
+  storage = clientStorage,
   history,
   store,
   context,
+  routes,
   renderRoot = null
 }, callback = () => null) => {
-
-
+  
   let component = (
     <Provider store={store}>
-      <ConnectedRouter Router={Router} context={context} history={history}>
-        <Switch>
-          <Route component={NotFoundPage} />
-        </Switch>
+      <ConnectedRouter
+        Router={Router}
+        context={context}
+        history={history}
+      >
+        <RootComponent
+          api={api}
+          storage={storage}
+          routes={routes}
+        >
+          <Switch>
+            <Route component={NotFoundPage} />
+          </Switch>
+        </RootComponent>
       </ConnectedRouter>
+    
     </Provider>
   );
   // If render is set false explicitly then just return the component
@@ -54,21 +73,78 @@ export const renderErrorPage = ({
   render = DefaultRender,
   Router = DefaultRouter,
   Switch = DefaultSwitch,
+  api = clientApi,
+  storage = clientStorage,
   context,
   history,
   store,
+  routes,
   renderRoot = null,
   error
 }, callback = () => null) => {
-
+  
   context = context || {};
   let component = (
     <HotAppContainer>
       <Provider store={store}>
-        <ConnectedRouter Router={Router} context={context} history={history} >
-          <Switch>
-            <ErrorPage error={error} />
-          </Switch>
+        <ConnectedRouter
+          Router={Router}
+          context={context}
+          history={history}
+        >
+          <RootComponent
+            api={api}
+            storage={storage}
+            routes={routes}
+          >
+            <Switch>
+              <ErrorPage error={error} />
+            </Switch>
+          </RootComponent>
+        </ConnectedRouter>
+      </Provider>
+    </HotAppContainer>
+  );
+  
+  if (!render) {
+    return component;
+  }
+  // Render 500
+  return render(component, renderRoot, callback);
+};
+
+export const renderOfflinePage = ({
+  render = DefaultRender,
+  Router = DefaultRouter,
+  Switch = DefaultSwitch,
+  api = clientApi,
+  storage = clientStorage,
+  context,
+  history,
+  store,
+  routes,
+  renderRoot = null,
+  error
+}, callback = () => null) => {
+  
+  context = context || {};
+  let component = (
+    <HotAppContainer>
+      <Provider store={store}>
+        <ConnectedRouter
+          Router={Router}
+          context={context}
+          history={history}
+        >
+          <RootComponent
+            api={api}
+            storage={storage}
+            routes={routes}
+          >
+            <Switch>
+              <OfflinePage error={error} />
+            </Switch>
+          </RootComponent>
         </ConnectedRouter>
       </Provider>
     </HotAppContainer>
@@ -95,10 +171,10 @@ export const renderRoutesByUrl = ({
   renderRoot = null
 }, callback = () => null) => {
   const currentRoutes = url ? getRouteFromPath(routes, url): routes;
-
+  
   context.api = api;
   context.storage = storage;
-
+  
   let component = (
     <HotAppContainer>
       <Provider store={store}>
@@ -108,18 +184,24 @@ export const renderRoutesByUrl = ({
           history={history}
           Router={Router}
         >
-          <Loader>
-            <Switch>
-              {_.map(currentRoutes, (route, i) => {
-                return <RouteWithSubRoutes
-                  key={i}
-                  route={route}
-                  storage={storage}
-                  api={api}
-                />;
-              })}
-            </Switch>
-          </Loader>
+          <RootComponent
+            api={api}
+            storage={storage}
+            routes={routes}
+          >
+            <Loader>
+              <Switch>
+                {_.map(currentRoutes, (route, i) => {
+                  return <RouteWithSubRoutes
+                    key={i}
+                    route={route}
+                    storage={storage}
+                    api={api}
+                  />;
+                })}
+              </Switch>
+            </Loader>
+          </RootComponent>
         </ConnectedRouter>
       </Provider>
     </HotAppContainer>
@@ -131,13 +213,15 @@ export const renderRoutesByUrl = ({
 };
 
 export const renderSubRoutes = (component) => {
-  const { routes = [], api, storage } = component.props;
+  const { routes = [], api, storage, preLoadedData } = component.props;
   return _.map(routes, (route, i) => {
     return <RouteWithSubRoutes
       key={i}
       route={route}
       api={api}
       storage={storage}
+      routes={routes}
+      parentPreLoadedData={preLoadedData}
     />;
   });
 };
@@ -147,6 +231,8 @@ export const renderSubRoutes = (component) => {
  * @param routes
  * @param storage
  * @param api
+ * @param store
+ * @param url
  * @returns {Array}
  */
 export const getPreloadDataPromises = (
@@ -154,18 +240,19 @@ export const getPreloadDataPromises = (
     routes,
     storage,
     api,
-    store
+    store,
+    url
   }
 ) => {
   let promises = [];
   _.each(routes, r => {
-
+    
     // Load data and add it to route itself
     if (r.preLoadData) {
       promises.push((() => {
-
+        
         // Pass route as reference so that we can modify it while loading data
-        let returnData = r.preLoadData({route: r, match: r.match, storage, api, store});
+        let returnData = r.preLoadData({route: r, match: r.match, storage, api, store, url});
         if (returnData && _.isFunction(returnData.then)) {
           return returnData.then(data => {
             return r.preLoadedData = data;
